requirements.txt
pandas
numpy
matplotlib
wordcloud
periodictable
seaborn
xmltodict
requests
nvcl_kit
fpdf2
pillow
